% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{barnsley}
\alias{barnsley}
\title{Barnsley fern}
\format{List of functions.}
\usage{
barnsley
}
\description{
This data set consist of functions set, which represent
Barnsley fern.
}
\keyword{datasets}
