% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{binary_tree}
\alias{binary_tree}
\title{Binary tree}
\format{List of functions.}
\usage{
binary_tree
}
\description{
This data set consist of functions set, which represent
binary tree fractal.
Taken from \url{http://paulbourke.net/fractals/}.
}
\keyword{datasets}
