% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{christmas}
\alias{christmas}
\title{Christmas tree}
\format{List of functions.}
\usage{
christmas
}
\description{
This data set consist of functions set, which represent
christmas tree fractal.
Taken from \url{http://paulbourke.net/fractals/}.
}
\keyword{datasets}
