% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFSanimation.R
\name{createIFSgif}
\alias{createIFSgif}
\title{Creating gifs for a given fractal}
\usage{
createIFSgif(startPoint, depth, fractal, resolution = 2, filename,
  scale = 1)
}
\arguments{
\item{startPoint}{from, which point (vertor of length 2) fractal is created}

\item{depth}{how depth is fractal}

\item{fractal}{the fractal from which the gif is made (list of function)}

\item{resolution}{how exactly do we paint a gif}

\item{filename}{the name with which the file is saved}

\item{scale}{how many times the gif is enlarged}
}
\description{
The \code{createIFSgif} function, writes a fractal with given parameters
on the disk
}
\examples{
f1 <- function(p){return(c(p[1]/2 - 0.5, p[2]/2 + 0.5))}
f2 <- function(p){return(c(p[1]/2 - 0.5, p[2]/2 - 0.5))}
f3 <- function(p){return(c(p[1]/2 + 0.5, p[2]/2 - 0.5))}

createIFSgif(startPoint = c(0,0), depth =8, list(f1,f2,f3),
            resolution = 2, filename = "output.gif", scale = 1)
}
