% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{dragon}
\alias{dragon}
\title{Dragon curve}
\format{List of functions.}
\usage{
dragon
}
\description{
This data set consist of functions set, which represent
Dragon curve.
Taken from \url{http://paulbourke.net/fractals/}.
}
\keyword{datasets}
