% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{heighway}
\alias{heighway}
\title{Heighway dragon}
\format{List of functions.}
\usage{
heighway
}
\description{
This data set consist of functions set, which represent
Heighway dragon.
}
\keyword{datasets}
