% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFSrainbow.R
\docType{package}
\name{IFSrainbow-package}
\alias{IFSrainbow}
\alias{IFSrainbow-package}
\title{Creating, plotting and basic manipulation of fractals}
\description{
\code{IFSrainbow} allows you to plot nice and highly customisable
pictures of IFS fractals. See \url{https://en.wikipedia.org/wiki/Iterated_function_system}
for more information about IFS fractals.
}
\details{
The only two functions that you are likely to use are \code{\link{createIFS}} (to
create an IFS fractal) and \code{\link{plot.IFS}} to plot it. We can also make you a GIF!
Moreover you may fancy adding
fractals with the use of an overloaded \code{\link{+.IFS}} operator.
}
\author{
\strong{Maintainer}: Jacek Maksymiuk \email{jm332326@students.mimuw.edu.pl}

Authors:
\itemize{
  \item Jacek Maksymiuk \email{jm332326@students.mimuw.edu.pl}
  \item Lukasz Rajkowski \email{l.m.rajkowski@gmail.com}
}

}
