% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{koch}
\alias{koch}
\title{Koch snowflake}
\format{List of functions.}
\usage{
koch
}
\description{
This data set consist of functions set, which represent
Koch snowflake.
}
\keyword{datasets}
