% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIFS.R
\name{plot.IFS}
\alias{plot.IFS}
\title{Plot a fractal}
\usage{
\method{plot}{IFS}(x, n, depth, probs = replicate(length(x), 1),
  inits = "random", seed = FALSE, ...)
}
\arguments{
\item{x}{an S3 object of class \code{IFS}, as returned by
\code{\link{createIFS}}}

\item{n}{number of points in the resulting plot}

\item{depth}{length of compositions of functions used to obtain
the plotted points}

\item{probs}{optional argument that defines probabilities of using
respective functions.
By defalt it is a sequence of 1's (i.e. all functions are equally probable).}

\item{inits}{optional argument being a matrix with \code{n} rows and
2 columns which defines a
set of \code{n} points. These points serve as arguments for the compositions of
length \code{depth}
of function given by the \code{ifs} argument. By default it is set to
\code{"random"},
which means that these points are sampled from the uniform distribution on [-1,1]
by [-1,1] square.}

\item{seed}{optional parameter that allows for setting the seed and hence makes
replication possible. The default
value is \code{FALSE} which means that the seed is not changed.}

\item{...}{graphical parameters passed to \code{\link[ggplot2]{ggplot}}
function}
}
\value{
The output is an object of class \code{\link[ggplot2]{ggplot}},
which allows for some further modifications
of the plot, as indicated in the Examples.
}
\description{
\code{plot.IFS} plots fractals obtained by \code{\link{createIFS}} function
}
\details{
This function plots IFS fractals. By default it uses the method
called "chaos game" in which random points are transformed by a
random composition of the fractal defining
functions.
}
\examples{
p1 <- plot(heighway, 2000, 10)
p1
p2 <- plot(heighway, 2000, 10, probs=c(2,1),
           inits=matrix(c(1,1), 1, 2), color="black")
p2
p1 + ggplot2::coord_fixed() + ggplot2::ggtitle("Heighway Dragon")

}
