% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusIFS.R
\name{+.IFS}
\alias{+.IFS}
\title{Addition of fractals}
\usage{
\method{+}{IFS}(x, y)
}
\arguments{
\item{x}{S3 objects of class \code{IFS}, as returned by}

\item{y}{S3 objects of class \code{IFS}, as returned by
\code{\link{createIFS}}}
}
\description{
\code{+.IFS} is an overloaded operator that allows to add two fractals
}
\details{
Adding two IFS fractals is just a fractal defined by the
union of functions defining the summands.
}
\examples{
newfractal <- heighway + koch
plot(newfractal, 100, 10)
}
