% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{real_tree}
\alias{real_tree}
\title{Tree}
\format{List of functions.}
\usage{
real_tree
}
\description{
This data set consist of functions set, which represent
tree fractal.
Taken from \url{http://paulbourke.net/fractals/}.
}
\keyword{datasets}
