% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{sierpinski}
\alias{sierpinski}
\title{Sierpinski triangle}
\format{List of functions.}
\usage{
sierpinski
}
\description{
This data set consist of functions set, which represent
Sierpinski triangle.
}
\keyword{datasets}
