% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetsDesc.R
\docType{data}
\name{spiral}
\alias{spiral}
\title{Spiral}
\format{List of functions.}
\usage{
spiral
}
\description{
This data set consist of functions set, which represent
spiral fractal.
Taken from \url{http://paulbourke.net/fractals/}.
}
\keyword{datasets}
